+++ src/database.py
@@ -6,5 +6,5 @@ def main():
 	client=pm.MongoClient('localhost', 27017) #connecting to mongodb
 	db=client['thinkit'] #connecting to database
 	offers=db.offers #getting collection
-	db.offers.delete_many({})
+	#db.offers.delete_many({})
 	return db
